Refactor: Show Basket updated

+++ AllUp2/AllUp2/Models/Product.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Price { get; set; }
-        public decimal DiscountPrice { get; set; }
+        public double Price { get; set; }
+        public double DiscountPrice { get; set; }
         public int Count { get; set; }
         public string ProductCode { get; set; }
         public bool IStock { get; set; }
